+++ packages/checkout/src/constants/abi.ts
@@ -183,84 +183,3 @@ export const ERC_1155_SALE_CONTRACT = [
     "stateMutability": "view"
   }
 ]
-
-export const ERC_721_SALE_CONTRACT = [
-  {
-    "type": "function",
-    "name": "mint",
-    "inputs": [
-      {
-        "name": "to",
-        "type": "address",
-        "internalType": "address"
-      },
-      {
-        "name": "amount",
-        "type": "uint256",
-        "internalType": "uint256"
-      },
-      {
-        "name": "paymentToken",
-        "type": "address",
-        "internalType": "address"
-      },
-      {
-        "name": "maxTotal",
-        "type": "uint256",
-        "internalType": "uint256"
-      },
-      {
-        "name": "proof",
-        "type": "bytes32[]",
-        "internalType": "bytes32[]"
-      }
-    ],
-    "outputs": [],
-    "stateMutability": "payable"
-  },
-  {
-    "type": "function",
-    "name": "saleDetails",
-    "inputs": [],
-    "outputs": [
-      {
-        "name": "",
-        "type": "tuple",
-        "internalType": "struct IERC721SaleFunctions.SaleDetails",
-        "components": [
-          {
-            "name": "supplyCap",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "cost",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "paymentToken",
-            "type": "address",
-            "internalType": "address"
-          },
-          {
-            "name": "startTime",
-            "type": "uint64",
-            "internalType": "uint64"
-          },
-          {
-            "name": "endTime",
-            "type": "uint64",
-            "internalType": "uint64"
-          },
-          {
-            "name": "merkleRoot",
-            "type": "bytes32",
-            "internalType": "bytes32"
-          }
-        ]
-      }
-    ],
-    "stateMutability": "view"
-  },
-]
